pyproject.toml
[tool.poetry]
name = "torchcu"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
together = "^1.2.12"
python-dotenv = "^1.0.1"
ruff = "^0.6.7"
flake8 = "^7.1.1"
torch = "^2.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
